Adding Full Text Search support to partitions
Contains Spiral patches:
- Spiral Patch 007: 007_mariadb-10.2.0.partition_fulltext.diff  MDEV-7705
- Spiral Patch 038: 038_mariadb-10.2.0.partition_fulltext2.diff MDEV-7734

- This commit has the following differences compared to the original patches:
  - Added necessary full text search cleanup at the storage engine layer
    that was omitted in the original patch.
  - Added test case.
- A lot of code cleanups to make the code notable smaller.
- Changed SQL code to use ha_ft_end() instead of ft_end()

Original author: Kentoku SHIBA
First reviewer:  Jacob Mathew
Second reviewer: Michael Widenius

+++ sql/handler.cc
@@ -1673,7 +1673,7 @@ int ha_rollback_trans(THD *thd, bool all)
         error=1;
 #ifdef WITH_WSREP
          WSREP_WARN("handlerton rollback failed, thd %lu %lld conf %d SQL %s",
-                    thd->thread_id, thd->query_id, thd->wsrep_conflict_state,
+                    (ulong) thd->thread_id, thd->query_id, thd->wsrep_conflict_state,
                     thd->query());
 #endif /* WITH_WSREP */
       }
